fix(vm): do not overwrite `cpu` attributes with defaults when cloning
